requirements.txt
scikit_image==0.18.2
face_recognition==1.3.0
Pillow==8.3.2
scikit_learn==0.24.2
skimage==0.0
